Use context.Context for control
Get rid of Stop() methods for most implementations, and instead
pass in a context.Context. Anything which needs to stop/cleanup
but doesn't care about shutdown order should wait on ctx.Done().

Also rework action stats to be tied to an agent instance
rather than a global variable. This allows for cleaner
startup/shutdown in test.