PATH-5 support file_id in action data
Add support for including both file_id and mover data in
in the action data field. This is currently assumed to be a
json with file_id and mover_data fields. Both fields
are opaque byte arrays, but are interpreted by the agent and
passed to the mover separately.

This change also now processes the file_id and data fields in
the initial thread that receives the action, instead of in the
request processing thread that is sending requests to the mover.