Refactor agent e2e test
This makes thje Agent more testable by integrating the starutp notification.

Tests now create the hsm.TestSource and an inject requests directly into the
source.  The idea is the test controls both of the ends of the pipeline
(TestSource and tewstMover), and everything inbetween (agent, dmplugin, etc)
are all created the same way they are normally. Hopefully.