PATH-3 Add compresion support to POSIX mover
Configuration is currently just on/off and not adaptive. That will
be worked on next.

To improve testability, I also consilidated all the configuration of
the mover into the posix directory, and updated NewMover to accept a
config struct. This allows test cases to change any config option in
ArchiveConfig, instead of being limited to ChecksumConfig.