LMR-4: Add workaround for radosgw support
In the short term, it is possible to work around the multipart
upload incompatibility by setting upload_part_size = 5000 (5GB).
The upshot with this setting is that the maximum file size that
can be archived is <= 5GB.

Longer-term, we need to figure out if we can fix the incompatibility
or at worst automatically fall back.